requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
distro==1.9.0
    # via openai
filelock==3.17.0
    # via huggingface-hub
fsspec==2025.2.0
    # via
    #   gradio-client
    #   huggingface-hub
google-auth==2.38.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.1
    # via gspread
gradio-client==1.7.2
    # via -r requirements.in
gspread==6.1.4
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   gradio-client
    #   openai
huggingface-hub==0.29.1
    # via gradio-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.8.2
    # via openai
loguru==0.7.3
    # via -r requirements.in
numpy==2.2.3
    # via -r requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.64.0
    # via -r requirements.in
packaging==24.2
    # via
    #   gradio-client
    #   huggingface-hub
pillow==11.1.0
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via huggingface-hub
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
websockets==15.0
    # via gradio-client
